+++ src/main/java/com/hrtx/web/mapper/OrderMapper.java
@@ -22,11 +22,11 @@ public interface OrderMapper extends Mapper<Order>,BaseMapper<Order>{
 
     List<Map> findOrderListByNumIdAndConsumerId(@Param("numId") Integer numId,@Param("consumerId") Integer consumerId);
 
-    Order findOrderInfo(@Param("orderid") Long id);
+    Order findOrderInfo(@Param("orderid") Integer id);
 
     Page<Object> getOrderByConsumer(@Param("order") Order order, @Param("st") String st);
 
-    Page<Object> queryPageList(@Param("param") Order order, @Param("sellerId") Long sellerId);
+    Page<Object> queryPageList(@Param("param") Order order, @Param("sellerId") Integer sellerId);
 
 
     /**----------------benging---取消订单操作涉及的方法--------------------------**/
@@ -38,11 +38,11 @@ public interface OrderMapper extends Mapper<Order>,BaseMapper<Order>{
      * @param status
      * @param reason
      */
-    void CancelOrderStatus(@Param("orderId") Long orderId, @Param("status") int status, @Param("reason") String reason);
+    void CancelOrderStatus(@Param("orderId") Integer orderId, @Param("status") int status, @Param("reason") String reason);
 
-    List getOrderItmeList(@Param("orderId") Long orderId,@Param("isShipment") int isShipment);
+    List getOrderItmeList(@Param("orderId") Integer orderId,@Param("isShipment") int isShipment);
 
-    List getOrderItmeCount(@Param("orderId") Long orderId,@Param("isShipment") int isShipment);
+    List getOrderItmeCount(@Param("orderId") Integer orderId,@Param("isShipment") int isShipment);
 
     /**
      *查询下单时间后两个小时未付款的订单
