+++ src/main/java/com/hrtx/web/pojo/Num.java
@@ -7,7 +7,7 @@ public class Num extends BasePojo{
     @Id
     private Integer id;
 
-    private Long cityId;
+    private Integer cityId;
 
     private String cityName;
 
@@ -37,9 +37,9 @@ public class Num extends BasePojo{
 
     private String buyer;
 
-    private Long iccidId;
+    private Integer iccidId;
     private String iccid;
-    private Long mealMid;
+    private Integer mealMid;
     private String slReason;
     private String uploadFileName;
     private Integer isFreeze;
@@ -47,9 +47,9 @@ public class Num extends BasePojo{
 
     private Integer status; //1在库、2销售中、3冻结(下单未付款)、4待配卡(已付款 针对2C或电销无需购买卡时、代理商买号而未指定白卡时)、5待受理(代理商已提交或仓库已发货，待提交乐语BOSS)、6已受理(乐语BOSS处理成功)、7受理失败(BOSS受理失败，需要人介入解决)、8已失效(乐语BOSS提示号码已非可用)
 
-    public Num(Integer id, Long cityId, String cityName, String netType, String numResource, String numType,
+    public Num(Integer id, Integer cityId, String cityName, String netType, String numResource, String numType,
                String numLevel, Double lowConsume, Boolean with4, String feature, String sectionNo, String moreDigit,
-               Integer sellerId, String seller, Integer buyerId, String buyer, String iccid, Integer status,Long mealMid) {
+               Integer sellerId, String seller, Integer buyerId, String buyer, String iccid, Integer status,Integer mealMid) {
         this.id = id;
         this.cityId = cityId;
         this.cityName = cityName;
@@ -83,11 +83,11 @@ public class Num extends BasePojo{
         this.id = id;
     }
 
-    public Long getCityId() {
+    public Integer getCityId() {
         return cityId;
     }
 
-    public void setCityId(Long cityId) {
+    public void setCityId(Integer cityId) {
         this.cityId = cityId;
     }
 
@@ -219,17 +219,17 @@ public class Num extends BasePojo{
         this.status = status;
     }
 
-    public Long getIccidId() {
+    public Integer getIccidId() {
         return iccidId;
     }
 
-    public void setIccidId(Long iccidId) {
+    public void setIccidId(Integer iccidId) {
         this.iccidId = iccidId;
     }
 
-    public Long getMealMid() { return mealMid; }
+    public Integer getMealMid() { return mealMid; }
 
-    public void setMealMid(Long mealMid) { this.mealMid = mealMid; }
+    public void setMealMid(Integer mealMid) { this.mealMid = mealMid; }
 
     public String getSlReason() {
         return slReason;
