+++ src/main/java/com/hrtx/web/service/DictService.java
@@ -30,7 +30,7 @@ public class DictService {
 		return new Result(Result.OK, pm);
 	}
 
-	public Dict findDictById(Long id) {
+	public Dict findDictById(Integer id) {
 		Dict dict = dictMapper.findDictInfo(id);
 		return dict;
 	}
